+++ lib/ts/recipe/dashboard/utils.ts
@@ -28,8 +28,6 @@ import AccountLinking from "../accountlinking/recipe";
 import EmailPasswordRecipe from "../emailpassword/recipe";
 import ThirdPartyRecipe from "../thirdparty/recipe";
 import PasswordlessRecipe from "../passwordless/recipe";
-import ThirdPartyEmailPasswordRecipe from "../thirdpartyemailpassword/recipe";
-import ThirdPartyPasswordlessRecipe from "../thirdpartypasswordless/recipe";
 import RecipeUserId from "../../recipeUserId";
 import { User, UserContext } from "../../types";
 import { logDebugMessage } from "../../logger";
@@ -73,13 +71,7 @@ export async function getUserForRecipeId(
     userContext: UserContext
 ): Promise<{
     user: UserWithFirstAndLastName | undefined;
-    recipe:
-        | "emailpassword"
-        | "thirdparty"
-        | "passwordless"
-        | "thirdpartyemailpassword"
-        | "thirdpartypasswordless"
-        | undefined;
+    recipe: "emailpassword" | "thirdparty" | "passwordless" | undefined;
 }> {
     let userResponse = await _getUserForRecipeId(recipeUserId, recipeId, userContext);
     let user: UserWithFirstAndLastName | undefined = undefined;
@@ -102,21 +94,9 @@ async function _getUserForRecipeId(
     userContext: UserContext
 ): Promise<{
     user: User | undefined;
-    recipe:
-        | "emailpassword"
-        | "thirdparty"
-        | "passwordless"
-        | "thirdpartyemailpassword"
-        | "thirdpartypasswordless"
-        | undefined;
+    recipe: "emailpassword" | "thirdparty" | "passwordless" | undefined;
 }> {
-    let recipe:
-        | "emailpassword"
-        | "thirdparty"
-        | "passwordless"
-        | "thirdpartyemailpassword"
-        | "thirdpartypasswordless"
-        | undefined;
+    let recipe: "emailpassword" | "thirdparty" | "passwordless" | undefined;
 
     const user = await AccountLinking.getInstance().recipeInterfaceImpl.getUser({
         userId: recipeUserId.getAsString(),
@@ -149,41 +129,12 @@ async function _getUserForRecipeId(
         } catch (e) {
             // No - op
         }
-
-        if (recipe === undefined) {
-            try {
-                // we detect if this recipe has been init or not..
-                ThirdPartyEmailPasswordRecipe.getInstanceOrThrowError();
-                recipe = "thirdpartyemailpassword";
-            } catch (e) {
-                // No - op
-            }
-        }
     } else if (recipeId === ThirdPartyRecipe.RECIPE_ID) {
         try {
             ThirdPartyRecipe.getInstanceOrThrowError();
             recipe = "thirdparty";
         } catch (e) {
             // No - op
-        }
-
-        if (recipe === undefined) {
-            try {
-                // we detect if this recipe has been init or not..
-                ThirdPartyEmailPasswordRecipe.getInstanceOrThrowError();
-                recipe = "thirdpartyemailpassword";
-            } catch (e) {
-                // No - op
-            }
-        }
-
-        if (recipe === undefined) {
-            try {
-                ThirdPartyPasswordlessRecipe.getInstanceOrThrowError();
-                recipe = "thirdpartypasswordless";
-            } catch (e) {
-                // No - op
-            }
         }
     } else if (recipeId === PasswordlessRecipe.RECIPE_ID) {
         try {
@@ -192,71 +143,11 @@ async function _getUserForRecipeId(
         } catch (e) {
             // No - op
         }
-
-        if (recipe === undefined) {
-            try {
-                ThirdPartyPasswordlessRecipe.getInstanceOrThrowError();
-                recipe = "thirdpartypasswordless";
-            } catch (e) {
-                // No - op
-            }
-        }
     }
     return {
         user,
         recipe,
     };
-}
-
-export function isRecipeInitialised(recipeId: RecipeIdForUser): boolean {
-    let isRecipeInitialised = false;
-
-    if (recipeId === "emailpassword") {
-        try {
-            EmailPasswordRecipe.getInstanceOrThrowError();
-            isRecipeInitialised = true;
-        } catch (_) {}
-
-        if (!isRecipeInitialised) {
-            try {
-                ThirdPartyEmailPasswordRecipe.getInstanceOrThrowError();
-                isRecipeInitialised = true;
-            } catch (_) {}
-        }
-    } else if (recipeId === "passwordless") {
-        try {
-            PasswordlessRecipe.getInstanceOrThrowError();
-            isRecipeInitialised = true;
-        } catch (_) {}
-
-        if (!isRecipeInitialised) {
-            try {
-                ThirdPartyPasswordlessRecipe.getInstanceOrThrowError();
-                isRecipeInitialised = true;
-            } catch (_) {}
-        }
-    } else if (recipeId === "thirdparty") {
-        try {
-            ThirdPartyRecipe.getInstanceOrThrowError();
-            isRecipeInitialised = true;
-        } catch (_) {}
-
-        if (!isRecipeInitialised) {
-            try {
-                ThirdPartyEmailPasswordRecipe.getInstanceOrThrowError();
-                isRecipeInitialised = true;
-            } catch (_) {}
-        }
-
-        if (!isRecipeInitialised) {
-            try {
-                ThirdPartyPasswordlessRecipe.getInstanceOrThrowError();
-                isRecipeInitialised = true;
-            } catch (_) {}
-        }
-    }
-
-    return isRecipeInitialised;
 }
 
 export async function validateApiKey(input: {